[api] extract page mod
